refactor(frontend): 拆分客户端大文件第二组
